requirements.txt
docopt==0.6.2
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
